AAAA

+++ tang-server/src/main/java/com/ooqn/repository/DtsArticleSubjectRelevanceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface DtsArticleSubjectRelevanceRepository extends CrudRepository<DtsSubjectRelevance, String> {
 
-    int countByArticleId(String articleId);
+    long countByArticleId(String articleId);
 
     DtsSubjectRelevance findByArticleId(String articleId);
